for all numbers

+++ 0x03-debugging/2-largest_number.c
@@ -10,7 +10,7 @@
 
 int largest_number(int a, int b, int c)
 {
-	int largest;
+	double largest;
 
 	if (a > b && b > c)
 	{
